claude/iTOP/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
requests==2.31.0
python-multipart==0.0.9
openpyxl==3.1.5
reportlab==4.4.10
